ai-coding-demo/requirements.txt
# Python 依赖

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
redis==5.0.1

# 安全
bcrypt==4.1.1
pyjwt==2.8.0
python-multipart==0.0.6

# 工具
python-dotenv==1.0.0
pyyaml==6.0.1
markdown==3.5.1

# 测试
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0

# 代码质量
black==23.12.1
flake8==7.0.0
pylint==3.0.3
mypy==1.7.1

# 开发工具
ipython==8.18.1
ipdb==0.13.13

# 监控和日志
structlog==23.2.0
prometheus-client==0.19.0
